drm/i915: Enable TDR / per-engine hang recovery
This is the final enablement patch for per-engine hang recovery. It sets up
per-engine hang recovery to be used per default in favour of full GPU reset.
Legacy full GPU reset will no longer be the preferred mode of hang recovery and
will only be used as a fall-back in case of frequent hangs on individual
engines or in the case of engine hang recovery failures.

Signed-off-by: Tomas Elf <tomas.elf@intel.com>